Add user stories to main

+++ src/com/skilldistillery/makechange/CashRegister.java
@@ -13,19 +13,33 @@ public class CashRegister {
 		price = sc.nextDouble();
 		System.out.println("Please enter the amount tendered: ");
 		tender = sc.nextDouble();
+
 		if (price > tender) {
 			System.out.println("Error: Come back when you get some more money buddy!");
-		}	else {
-			
+		} else {
+
 		}
 		sc.close();
 	}
 
 	private static double change(double price, double tender) {
 		double change = 0;
-		
-		return (double)change;
+
+		return change;
 
 	}
 
 }
+
+//	User Story #1
+//	The user is prompted asking for the price of the item. CW
+//	
+//	User Story #2
+//	The user is then prompted asking how much money was tendered by the customer.
+//	
+//	User Story #3
+//	Display an appropriate message if the customer provided too little money or the exact amount.
+//	
+//	User Story #4
+//	If the amount tendered is more than the cost of the item, display the number 
+//	of bills and coins that should be given to the customer.
